requirements.txt
pymongo==3.11.0
python-dotenv==1.0.0
python-telegram-bot==20.4
Unidecode==1.3.6
